git-svn-id: https://svn.eduke32.com/eduke32@610 1a8010ca-5511-0410-912e-c29ae57300e0

+++ polymer/eduke32/source/gamedef.h
@@ -41,7 +41,7 @@ extern int *actorLoadEventScrptr[MAXTILES];
 
 extern int *apScriptGameEvent[MAXGAMEEVENTS];
 
-extern short otherp;
+extern int otherp;
 extern int g_currentweapon;
 extern int g_gun_pos;
 extern int g_looking_arc;
